backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.9
httpx==0.27.2
sqlalchemy==2.0.36
databases[sqlite]==0.9.0
faiss-cpu==1.8.0
sentence-transformers==3.1.1
tiktoken==0.7.0
pymupdf==1.24.10
python-docx==1.1.2
tenacity==8.5.0
python-dotenv==1.0.1
orjson==3.10.7
